Fixed some text display stuff
Don't break Access codes into groups of 4 if they aren't a multiple of 4 or if they aren't hexadecimal only (This is useful if a user only uses a nickname as an access code and not a card id)

Changed color of is-current-user on the leaderboard so text is legible in dark mode.

Made Leaderboard useable on mobile as well by removing responsive breakpoints.