Update event_folder_data.json
The file was using Omnimix IDs causing empty entries on clean dumps of the game.

The consensus is the server should be working perfectly with clean copies and not modified versions so the new file reflects the official 39.06 folders. Many thanks to @CottonGuard who watched a ton of youtube stream replays to fill those folders properly!